Using systemd to run scylla-jmx won't load params from /etc/defaults/scylla-jmx  because it expects params to be sent using the command line.
This patch enhances the sysconfig file by adding the available options (commented) and passes the right options to systemd `ExecStart` when defined. That way scylla-jmx is ran with the params defined into /etc/defaults/scylla-jmx.

Maybe not the most elegant way to do it, but 1/ it works 2/ I didn't find a better solution to fix that problem.