install.sh: detect java located under /usr/lib/jvm also
both fedora/centos and debian/ubuntu use a system to configure
the symbolic link of /usr/bin/java. the former uses `alternatives`
while the latter uses `update-alternatives`. if user installs multiple
packages provides the same tool, the system picks one using some
grading system. but the package provides java 8 or java 11 does not
always win, so there are good chances that `install.sh` either
fails to detect the java executable or it leaves us a symbolic link
which points to /usr/bin/java, which could be reassigned to
another java provided by a new JRE. and the new java might not be
compatible with JMX.

in this change, we always try to detect the java using the
packaging system, and use it if it is available and fall back
to /usr/bin/java.

this should address the issues, we recently ran into after
upgrading a fedora 37 system. the symptom was scylla-jmx
exited immediately after upgrading the system.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>