build: install JRE instead devel package
java-1.8.0-openjdk-devel provides the OpenJDK 8 development environment,
in addition to the headless part, it also depends on the packages for
development in a GUI environment, like libXft and gtk2. but we don't
need them for building a maven project. what maven needs is but a JDK.
so, let's install java-1.8.0-openjdk-headless instead. this package
is much smaller than java-1.8.0-openjdk-devel.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>