install.sh: select java executable at runtime
before this change, a symbolic link to java runtime is created
in the install directory of scylla-jmx, and scylla-jmx uses it
to launch scylla-jmx-1.0.jar. and we only check for the expected
java version when `install.sh` is called without `--packaging`
specified. but, another important use case of `install.sh` is
for creating relocatable package. in that case, the `--packaging`
parameter is specified. so we always end up using `/usr/bin/java`
for launching jmx jar if jmx is installed from the relocatable
package.

but we do intend to use OpenJDK-11 for running scylla-jmx if
it is available, no matter how scylla-jmx is installed.

so, in this change, instead of relying on user pointing
`/usr/bin/java` to the appropriate java runtime, we just
choose it by ourselves. to accomplish this goal,
`scripts/select-java` is extracted out of `install.sh`.
this script picks JRE and execv it, and fails if neither
JRE-8 nor JRE-11 is found. to minimize the impact to other
parts of this project, while preserve a meaningful name
in the source tree, the name `select-java` is used, while
it is installed using the old name of `symlinks/scylla-jmx`,
even it is not a symblink anymore. also, instead of being
created by `install.sh`, `script/select-java` is included by
the relocatable package.

also, the error mesage is adjusted to place "openjdk-11" before
"openjdk-8" as a hint that openjdk-11 is prefered. this error
mesage is only printed out in non-packaging mode.

Fixes #194
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>