treewide: build jar which targets jdk-11
since we've moved to OpenJDK-11, let's build the jar targeting JDK-11.

the `--add-exports` change was created after referencing
https://docs.oracle.com/en/java/javase/16/migrate/migrating-jdk-8-later-jdk-releases.html#GUID-2F61F3A9-0979-46A4-8B49-325BA0EE8B66
without this change, we'd have build failures like
```
[ERROR] /home/kefu/dev/scylladb/tools/jmx/src/main/java/com/scylladb/jmx/utils/APIBuilder.java:[48,19] package com.sun.jmx.interceptor is not visible
[ERROR]   (package com.sun.jmx.interceptor is declared in module java.management, which does not export it)
```

so that the created jar file won't be compatible with JDK-8, this
helps us to identify the places where the JDK-8 is still used.

by passing `-source` to the javac compiler allows us to use the
Java-11 features. despite that we don't have plan to use any of
them, it is still a nice to have.

Refs #206
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>