dist/debian: drop dependencies on jdk-8
this change mirrors f89670dfd80ed04860e1098f0fc6ef114793f31b,
since the oldest supported debian and its derivative are debian 10 and
ubuntu 18.04, both of which ship openjdk-11-jre-headless. and
oracle provides oracle-java11-set-default for them. so it's safe
to drop the optional dependencies on openjdk-8 and orcale-java8.

this would ensure that the scylla-jmx deb package's access to
jdk-11.

Fixes #194
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>