install.sh: bail out if jdk is not available
in 96adf0ef51ad49b6abb98867f73cf69705f5f3e3, we started using
`select-java` for selecting the JRE for running scylla-jmx, but
it failed to reference `select-java` with the correct path. as
in the unified reloc package, `select-java` is located at
`scylla-jmx/select-java`, `install.sh` is located under the same
directory. so we should use `./select-java` here.

also, we should error out if the expected JDK is not found. this
is the behavior before 96adf0ef51ad49b6abb98867f73cf69705f5f3e3

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>