sstableinfo: change the type of generation to string
we will allow user to use UUID for the generation (identifier) of a
SSTable. So to be compatible to both "long" and an "UUID", let's use
"String" for the type of generation. this change requires the change
on scylladb side, which should expose the generation as a string also
in its "/storage_service/sstable_info" RESTful API.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>