scylla-apiclient: update Jackson dependencies
Update the version of Jackson dependencies to 2.15.3. Before the change,
security scanners (such as Trivy) reported that 
com.fasterxml.jackson.core:jackson-databind used in the project was
vulnerable to CVE-2022-42003 and CVE-2022-42004 ("HIGH" severity).

The issues are fixed in later versions of Jackson and after this
commit the security scanner doesn't report any problems related to
Jackson dependencies.

Fixes #221