scylla-apiclient: update Guava dependency
Update the version of Guava dependency to 32.1.3-jre. Before the change,
security scanners (such as Trivy) reported that com.google.guava used 
in the project was vulnerable to CVE-2023-2976 ("HIGH" severity) and 
CVE-2020-8908.

The issues are fixed in later versions of Guava and after this
commit the security scanner doesn't report any problems related to
Guava dependency.

Fixes #223