ColumnFamilyStore: don't fail if there is a table with ":" in its name
ColumnFamilyStore builds for its own uses a list of Scylla's tables using
the REST API. It holds each table's name in a formant meant for MBean,
in an ObjectName object. As the ObjectName documentation explains, when
a value contains a comma, equals, colon or quote - it must be quoted,
otherwise the ObjectName cannot be constructed.

Scylla table names can't ever contain commas, equals or quotes - but
they can contain colons, in the case of Alternator GSIs as discovered
in issue #226. So in this patch we quote the columnfamily value (using
the ObjectName.quote()) function, to allow it to contain a colon.

Our code also wants to read back the ObjectName, so we unquote() it
to restore the unquoted name.

We could have quoted the keyspace name as well, not just the columnfamily
name, but this isn't necessary because even in Alternator, keyspace names
cannot contain colons (or commas or equals or quotes).

After this patch, "notetool info" starts to work even if an Alternator
GSI exists. "nodetool info" doesn't need to access any specific table,
but it still causes our JMX implementation to build that list of tables
which failed before this patch and now works.

Fixes #226.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>