APIBuilder: allow quoted scope names
After the previous patch, our JMX implementation no longer fails when
there is a table whose name contains the colon character (as in #226).
However, the JMX client (like nodetool) still can't *address* this
table - in order to be used in a MBean object name it must be quoted
(as explained in the documentation of ObjectName), but our server
doesn't understand the quoted name.

So in this patch we check if the "scope" (i.e., columnfamily name)
is quoted, and if it is, unquote it while searching for the table.
We don't do the same treatment for the other components of the bean
name (like keyspace) because those can't have a colon character.

After this patch, "nodetool tablestats" almost begins to work when
there are Alternator GSIs (with a colon in the table's name). It's
"almost" because we also need a patch in nodetool, so it will quote
the table name before sending it. The patch *here* is what will allow
this quoted name to be understood by the JMX server.

Refs #226.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>